Generate an event ID in `Batch` step
